Remove module list
12fd07a8af663423599bee2713aec3a18c6a0e77 removed the Travis secret
scripts, but left this list of modules. It doesn't really serve a
purpose anymore. If you want a complete list of modules, Github can list
it. modulesync_config should also have a list of actively maintained
modules with modulesync. While that leaves out archived and PDK modules,
it's not really a problem.